ready for initial plots
