+++ py/obiwan/priors_MoG.py
@@ -1,0 +1,171 @@
+
+class GaussianMixtureModel(object):
+    """Sample from a Mixture of Gaussians (MoG) with known params
+    
+	Args:
+		weights_,means_,covars_: MoG parameters
+		covtype: 'spherical', 'diag', 'tied', 'full'
+
+	Attributes:
+		weights_,means_,covars_: MoG parameters
+		covtype: 'spherical', 'diag', 'tied', 'full'
+		n_components: as sounds
+		n_dimenstions: as sounds
+    """
+	
+    def __init__(self, weights_, means_, covars_, covtype):
+        self.weights_ = weights_
+        self.means_ = means_
+        self.covars_ = covars_
+        self.covtype = covtype
+        self.n_components, self.n_dimensions = self.means_.shape
+    
+    @staticmethod
+    def save(model, filename,index=None):
+        """index: optional nddex array for subset of compenents to save"""
+        hdus = fits.HDUList()
+        hdr = fits.Header()
+        hdr['covtype'] = model.covariance_type
+        if index is None:
+            index=np.arange(len(model.weights_))
+        hdus.append(fits.ImageHDU(model.weights_[index], name='weights_', header=hdr))
+        hdus.append(fits.ImageHDU(model.means_[index,...], name='means_'))
+        hdus.append(fits.ImageHDU(model.covars_[index,...], name='covars_'))
+        hdus.writeto(filename, clobber=True)
+        
+    @staticmethod
+    def load(filename):
+        hdus = fits.open(filename, memmap=False)
+        hdr = hdus[0].header
+        covtype = hdr['covtype']
+        model = _GaussianMixtureModel(
+            hdus['weights_'].data, hdus['means_'].data, hdus['covars_'].data, covtype)
+        hdus.close()
+        return model
+    
+    def sample(self, n_samples=1, random_state=None):
+        
+        if self.covtype != 'full':
+            return NotImplementedError(
+                'covariance type "{0}" not implemented yet.'.format(self.covtype))
+        
+        # Code adapted from sklearn's GMM.sample()
+        if random_state is None:
+            random_state = np.random.RandomState()
+
+        weight_cdf = np.cumsum(self.weights_)
+        X = np.empty((n_samples, self.n_dimensions))
+        rand = random_state.rand(n_samples)
+        # decide which component to use for each sample
+        comps = weight_cdf.searchsorted(rand)
+        # for each component, generate all needed samples
+        for comp in range(self.n_components):
+            # occurrences of current component in X
+            comp_in_X = (comp == comps)
+            # number of those occurrences
+            num_comp_in_X = comp_in_X.sum()
+            if num_comp_in_X > 0:
+                X[comp_in_X] = random_state.multivariate_normal(
+                    self.means_[comp], self.covars_[comp], num_comp_in_X)
+        return X
+    
+    def sample_full_pdf(self, n_samples=1, random_state=None):
+        """'sample() uses the component datum x is closest too, sample_full_pdf() uses sum of components at datum x
+        this is more time consuming than sample() and difference is negligible"""
+        if self.covtype != 'full':
+            return NotImplementedError(
+                'covariance type "{0}" not implemented yet.'.format(self.covtype))
+
+        from scipy.stats import multivariate_normal
+        def get_mv(means_,covars_):
+            mv=[]
+            for mu, C in zip(means_, covars_):
+                mv+= [ multivariate_normal(mean=mu, cov=C) ]
+            return mv
+                
+        def prob_map(means_,covars_,weights_,\
+                     xrng=(0.,1.),yrng=(0.,1.),npts=2**10):
+            """returns 
+            -pmap: 2d probability map, with requirement that integral(pdf d2x) within 1% of 1
+            -xvec,yvec: vectors where x[ix] and y[ix] data points have probability pmap[ix,iy]"""
+            assert(xrng[1] > xrng[0] and yrng[1] > yrng[0])
+            xstep= (xrng[1]-xrng[0])/float(npts-1)
+            ystep= (yrng[1]-yrng[0])/float(npts-1)
+            x,y  = np.mgrid[xrng[0]:xrng[1]+xstep:xstep, yrng[0]:yrng[1]+ystep:ystep]
+            pos = np.empty(x.shape + (2,)) #npts x npts x 2
+            pos[:, :, 0] = x; pos[:, :, 1] = y
+            maps= np.zeros(x.shape+ (len(weights_),)) # x n_components
+            # Multi-Variate function
+            mv= get_mv(means_,covars_)
+            # probability map for each component
+            for dist,W,cnt in zip(mv,weights_, range(len(weights_))):
+                maps[:,:,cnt]= dist.pdf(pos) * W
+                print "map %d, dist.pdf max=%.2f, wt=%.3f" % (cnt,dist.pdf(pos).max(),W)
+            # summed prob map
+            pmap= np.sum(maps, axis=2) #some over components*weights
+            xvec= x[:,0]
+            yvec= y[0,:]
+            # intregal of pdf over 2d map = 1
+        #     assert( abs(1.- pmap.sum()*xstep*ystep) <= 0.01 )
+            assert( np.diff(xvec).min() > 0. and np.diff(yvec).min() > 0.)
+            return pmap, xvec,yvec
+            
+        # 2D probability map
+        grrange = (-0.2, 2.0)
+        rzrange = (-0.4, 2.5)
+        pmap,xvec,yvec= prob_map(self.means_,self.covars_,self.weights_,\
+                                 xrng=rzrange,yrng=grrange)
+        # Sample self.n_dimensions using map
+        if random_state is None:
+            r = np.random.RandomState()
+        # Make max pdf = 1 so always pick that cell
+        pmap/= pmap.max()
+        # Store x,y values to use
+        X = np.empty((n_samples, self.n_dimensions))+np.nan
+        # Get samples
+        cnt=0
+        # pick a random cell
+        ix,iy=(r.rand(2)*len(xvec)).astype(int)
+        # Random [0,1)
+        likely= r.rand(1)
+        while cnt < n_samples:
+            if likely <= pmap[ix,iy]: # Sample it!
+                X[cnt,:]= xvec[ix],yvec[iy]
+                cnt+=1
+            # Pick new cell in either case
+            ix,iy=(r.rand(2)*len(xvec)).astype(int) 
+            likely= r.rand(1)
+        assert( np.where(~np.isfinite(X))[0].size == 0)
+        return X
+
+
+def add_MoG_curves(ax, means_, covars_, weights_):
+    """plot 2-sigma ellipses for each multivariate component
+
+	Args: 
+		ax: axis to draw on
+		means_, covars_, weights_: MoG parameters
+	"""
+    ax.scatter(means_[:, 0], means_[:, 1], c='w')
+    scale=2.
+    for cnt, mu, C, w in zip(range(means_.shape[0]),means_, covars_, weights_):
+    #     draw_ellipse(mu, C, scales=[1.5], ax=ax, fc='none', ec='k')
+        # Draw MoG outlines
+        sigma_x2 = C[0, 0]
+        sigma_y2 = C[1, 1]
+        sigma_xy = C[0, 1]
+
+        alpha = 0.5 * np.arctan2(2 * sigma_xy,
+                             (sigma_x2 - sigma_y2))
+        tmp1 = 0.5 * (sigma_x2 + sigma_y2)
+        tmp2 = np.sqrt(0.25 * (sigma_x2 - sigma_y2) ** 2 + sigma_xy ** 2)
+
+        sigma1 = np.sqrt(tmp1 + tmp2)
+        sigma2 = np.sqrt(tmp1 - tmp2)
+        print('comp=%d, sigma1=%f,sigma2=%f' % (cnt+1,sigma1,sigma2))
+
+        ax.text(mu[0],mu[1],str(cnt+1),color='blue')
+        ax.add_patch(Ellipse((mu[0], mu[1]),
+                     2 * scale * sigma1, 2 * scale * sigma2,
+                     alpha * 180. / np.pi,\
+                     fc='none', ec='k'))
